Option to use audit.log to map usid -> username
The progress trace only contains MAAPI user session IDs. When tracing
events, it is more useful to also know the username.
A new option '-a <path/to/audit.log>' is added to the script. If the log
is provided, it is used to build a mapping of usid -> username. For
known usids, a new tag 'username' is added to the span.